reorg the tos data to match the new structure
