+++ land-ocean-warming-ratio/.idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,15 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="2">
+            <item index="0" class="java.lang.String" itemvalue="data.time.min" />
+            <item index="1" class="java.lang.String" itemvalue="data.coords" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
