+++ land-ocean-warming-ratio/README.md
@@ -1,0 +1,5 @@
+# land-ocean-warming-ratio
+Code to parameterize the land-ocean warming ration within the HECTOR simple climate model using CMIP6 data.
+Prepared by Skylar Gering as part of their 2020 SULI work, the original repo can be found here, https://github.com/skygering/land-ocean-warming-ratio but copied here so that all of Hector’s CMIP6 related code is consolidated in a single place. 
+
+Note this code was not written using hector_cmip6data as its root directory.
