+++ land-ocean-warming-ratio/bibliography.bib
@@ -1,0 +1,10 @@
+@article{Hector,
+  title = {A simple object-oriented and open-source model for scienti???c and policy analyses of the global climate system - Hector v1.0},
+  author = {{Hartin, C.A., Patel, P., Schwarber, A., 
+  Link, R.P., Bond-Lamberty, B.P.}},
+  organization = {Geoscientific Model Development (Online) 8},
+  number = {8}
+  issn = {PNNL-SA-104722},
+  year = {2015},
+  url = {www.geosci-model-dev.net/8/939/2015/}
+}
