+++ land-ocean-warming-ratio/ratio_cleaned_1pctCO2_temp.csv
@@ -1,0 +1,69 @@
+"Model","Ratio","Error","Year"
+"ACCESS-CM2",1.54188052399188,0.171065747562649,45
+"ACCESS-CM2",1.49761172097776,0.0940414334653467,75
+"ACCESS-CM2",1.50550002119171,0.0586300428134911,105
+"ACCESS-CM2",1.50692513542607,0.0444287771762664,135
+"ACCESS-ESM1-5",1.60075855064675,0.19993210605333,45
+"ACCESS-ESM1-5",1.60454145842767,0.100002072013217,75
+"ACCESS-ESM1-5",1.58174505630411,0.0626580604556684,105
+"ACCESS-ESM1-5",1.5638488469959,0.0472790133999952,135
+"BCC-ESM1",1.83612981891733,0.259339611761894,45
+"BCC-ESM1",1.62115083216292,0.0950565134773634,75
+"BCC-ESM1",1.65381337473802,0.0633546959338544,105
+"BCC-ESM1",1.67609746664103,0.0481029066689214,135
+"CanESM5",1.54980428617357,0.163742244054722,45
+"CanESM5",1.56241156026071,0.0761737066057282,75
+"CanESM5",1.54084345987086,0.0577946665763531,105
+"CanESM5",1.55161595962017,0.039603455597571,135
+"CESM2",1.45023253394145,0.159984661902707,45
+"CESM2",1.42192551762748,0.0754765646892447,75
+"CESM2",1.39004679685507,0.0597689283637863,105
+"CESM2",1.40373095334731,0.044635721960714,135
+"CESM2-WACCM",1.47201763715581,0.172681385262275,45
+"CESM2-WACCM",1.3929364686893,0.0880679188221422,75
+"CESM2-WACCM",1.40138686526705,0.0599697148431281,105
+"CESM2-WACCM",1.40418770287162,0.0468625096914383,135
+"E3SM-1-0",1.53510523579388,0.196032081460154,45
+"E3SM-1-0",1.49723178614512,0.0746651699257076,75
+"E3SM-1-0",1.48736764009914,0.0574800191747401,105
+"E3SM-1-0",1.50360854399117,0.0405975794352698,135
+"GFDL-CM4",1.67241427627776,0.16926477787065,45
+"GFDL-CM4",1.66609538840268,0.0950113771629733,75
+"GFDL-CM4",1.64873291692378,0.0679768436526085,105
+"GFDL-CM4",1.626948542759,0.0443661455764957,135
+"INM-CM4-8",1.44879415425101,0.149588518789643,45
+"INM-CM4-8",1.49754670792609,0.0920803015272733,75
+"INM-CM4-8",1.4494164462614,0.0556755370212547,105
+"INM-CM4-8",1.41748048945908,0.0388259095766643,135
+"IPSL-CM6A-LR",1.65794930073995,0.24386613452656,45
+"IPSL-CM6A-LR",1.59993443503227,0.103651363360823,75
+"IPSL-CM6A-LR",1.61262502465275,0.0659334993319708,105
+"IPSL-CM6A-LR",1.60766331717696,0.056299387874337,135
+"MIROC6",2.00265819980778,0.244730422262321,45
+"MIROC6",1.87052741558075,0.120873593249755,75
+"MIROC6",1.85143244545899,0.0763179359903984,105
+"MIROC6",1.82442378301543,0.0542216097477449,135
+"MPI-ESM1-2-LR",1.6335363700121,0.199362679593122,45
+"MPI-ESM1-2-LR",1.67814815009808,0.0996624421194674,75
+"MPI-ESM1-2-LR",1.64276323857076,0.0627063854809851,105
+"MPI-ESM1-2-LR",1.62900469343099,0.0510194955516173,135
+"MRI-ESM2-0",1.53262539834314,0.209508424376036,45
+"MRI-ESM2-0",1.37365405761974,0.0953969623729141,75
+"MRI-ESM2-0",1.33993378803977,0.0567402897793521,105
+"MRI-ESM2-0",1.3265966661907,0.0487674751528188,135
+"NorCPM1",1.67392585812326,0.217953232316195,45
+"NorCPM1",1.65394754882775,0.1051539838253,75
+"NorCPM1",1.65115796496428,0.0646252475023933,105
+"NorCPM1",1.61345875358367,0.0439958946509651,135
+"NorESM2-LM",1.74471634637498,0.224604603309287,45
+"NorESM2-LM",1.78246895501519,0.106989202867906,75
+"NorESM2-LM",1.79392622186964,0.0796550005755978,105
+"NorESM2-LM",1.78978721382174,0.0625378580552816,135
+"NorESM2-MM",1.68506717059822,0.240436808412827,45
+"NorESM2-MM",1.77607948382375,0.137976828177871,75
+"NorESM2-MM",1.76293593291859,0.0853117721638171,105
+"NorESM2-MM",1.71001704080369,0.0582946688348328,135
+"SAM0-UNICON",1.56015325583787,0.182834856000123,45
+"SAM0-UNICON",1.52578242123483,0.0833890691986007,75
+"SAM0-UNICON",1.47801113861964,0.0599459935403155,105
+"SAM0-UNICON",1.46500877565619,0.0405671999530349,135
